+++ packages/styles/bases/date-picker-input.css.ts
@@ -1,13 +1,15 @@
 import { style, styleVariants } from "@vanilla-extract/css";
 import { THEME } from "@wizleap-inc/wiz-ui-constants";
 
+const BORDER_WIDTH = "1px";
+
 export const datePickerStyle = style({
   width: "max-content",
-  height: THEME.spacing.xl3,
   borderRadius: THEME.spacing.xs2,
   boxSizing: "border-box",
-  padding: `${THEME.spacing.no} ${THEME.spacing.xs}`,
+  padding: `calc(${THEME.spacing.xs} - ${BORDER_WIDTH}) ${THEME.spacing.xs}`,
   fontSize: THEME.fontSize.sm,
+  lineHeight: THEME.fontSize.xl3,
 });
 
 export const datePickerVariantStyle = styleVariants({
@@ -22,7 +24,7 @@ export const datePickerVariantStyle = styleVariants({
     cursor: "pointer",
   },
   disabled: {
-    color: THEME.color.gray["700"],
+    color: THEME.color.gray["500"],
     background: THEME.color.gray["300"],
     cursor: "not-allowed",
   },
@@ -88,4 +90,16 @@ export const datePickerYearSelectorItemStyle = style({
 export const datePickerArrowIconStyle = style({
   transform: "scale(2)",
   pointerEvents: "none",
+});
+
+export const datePickerCancelButtonStyle = style({
+  cursor: "pointer",
+  padding: THEME.spacing.no,
+  borderRadius: THEME.spacing.xs2,
+  border: "none",
+  background: "transparent",
+  fill: THEME.color.gray["500"],
+  ":active": {
+    fill: THEME.color.green["800"],
+  },
 });