+++ packages/wiz-ui-next/src/components/base/inputs/date-picker/date-picker.vue
@@ -10,14 +10,36 @@
         width,
       }"
       :aria-label="ARIA_LABELS.DATE_PICKER_INPUT"
-      @click="toggleDatepicker"
+      @click="setIsOpen(!isOpen)"
     >
       <WizHStack gap="xs" align="center" height="100%">
-        <WizIcon size="xl2" color="gray.500" :icon="WizICalendar" />
-        <span>{{ parseValue(calendarValue) || placeholder }}</span>
+        <span
+          v-if="!isHover"
+          @mouseenter="setIsHover(true)"
+          @mouseleave="setIsHover(false)"
+        >
+          <WizIcon size="xl2" color="gray.500" :icon="WizICalendar" />
+        </span>
+        <button
+          v-else
+          :class="datePickerCancelButtonStyle"
+          :aria-label="ARIA_LABELS.DATE_PICKER_CANCEL"
+          @click="onClickCancel"
+          @mouseenter="setIsHover(true)"
+          @mouseleave="setIsHover(false)"
+        >
+          <WizIcon size="xl2" color="inherit" :icon="WizICancel" />
+        </button>
+        <span>{{
+          (calendarValue && formatDateToYYMMDD(calendarValue)) || placeholder
+        }}</span>
       </WizHStack>
     </button>
-    <WizPopup :isOpen="openDatepicker" @onClose="openDatepicker = false">
+    <WizPopup
+      :isOpen="!disabled && isOpen"
+      @onClose="setIsOpen(false)"
+      :isDirectionFixed="isDirectionFixed"
+    >
       <div :class="datePickerSelectorStyle">
         <WizHStack align="center" my="xs2" px="xs" justify="between">
           <WizHStack align="center" justify="between" gap="xs2">
@@ -96,41 +118,46 @@
 <script setup lang="ts">
 import { ARIA_LABELS } from "@wizleap-inc/wiz-ui-constants";
 import {
+  datePickerArrowIconStyle,
+  datePickerCancelButtonStyle,
+  datePickerMonthSelectorItemStyle,
+  datePickerMonthSelectorStyle,
+  datePickerSelectorStyle,
   datePickerStyle,
   datePickerVariantStyle,
-  datePickerSelectorStyle,
-  datePickerMonthSelectorStyle,
-  datePickerMonthSelectorItemStyle,
   datePickerYearSelectorItemStyle,
-  datePickerArrowIconStyle,
 } from "@wizleap-inc/wiz-ui-styles/bases/date-picker-input.css";
 import {
-  inputBorderStyle,
   fillStyle,
   fontSizeStyle,
+  inputBorderStyle,
 } from "@wizleap-inc/wiz-ui-styles/commons";
-import { ref, computed, inject, PropType } from "vue";
+import { formatDateToYYMMDD } from "@wizleap-inc/wiz-ui-utils";
+import { PropType, computed, inject, ref } from "vue";
 
 import {
-  WizIcon,
-  WizVStack,
-  WizHStack,
   WizCalendar,
-  WizText,
+  WizHStack,
+  WizIcon,
   WizPopup,
   WizPopupContainer,
+  WizText,
+  WizVStack,
 } from "@/components";
 import {
+  WizIArrowDropDown,
+  WizIArrowDropUp,
   WizICalendar,
+  WizICancel,
   WizIChevronLeft,
   WizIChevronRight,
-  WizIArrowDropUp,
-  WizIArrowDropDown,
 } from "@/components/icons";
 import { formControlKey } from "@/hooks/use-form-control-provider";
 
 interface Emit {
-  (e: "update:modelValue", value: Date): void;
+  (e: "update:modelValue", value: Date | null): void;
+  (e: "update:isOpen", value: boolean): void;
+  (e: "update:isHover", value: boolean): void;
 }
 
 const props = defineProps({
@@ -153,20 +180,35 @@ const props = defineProps({
     required: false,
     default: false,
   },
+  /**
+   * カレンダー（Popup）の開閉状態を指定します。
+   */
+  isOpen: {
+    type: Boolean,
+    required: true,
+  },
+  /**
+   * `isHover=true`の時、キャンセルアイコンを緑色にします。
+   */
+  isHover: {
+    type: Boolean,
+    required: true,
+  },
+  isDirectionFixed: {
+    type: Boolean,
+    required: false,
+    default: false,
+  },
 });
 
 const emit = defineEmits<Emit>();
 
-const defaultCurrentMonth = new Date().setHours(0, 0, 0, 0);
-const currentMonth = ref(new Date(defaultCurrentMonth));
-const openDatepicker = ref(false);
+const defaultCurrentMonth = props.modelValue || new Date();
+const currentMonth = ref(defaultCurrentMonth);
 
-const toggleDatepicker = () => {
-  if (props.disabled) {
-    return;
-  }
-  openDatepicker.value = !openDatepicker.value;
-};
+const setIsOpen = (value: boolean) => emit("update:isOpen", value);
+const setIsHover = (value: boolean) => emit("update:isHover", value);
+const onClickCancel = () => emit("update:modelValue", null);
 
 const clickToNextMonth = () => {
   const setDateTime = new Date(
@@ -204,12 +246,6 @@ const clickToPreviousYear = () => {
   currentMonth.value = new Date(setDateTime);
 };
 
-const parseValue = (inputValue: Date | null) => {
-  if (inputValue === null) return undefined;
-  const value = new Date(inputValue);
-  return `${value.getFullYear()}/${value.getMonth() + 1}/${value.getDate()}`;
-};
-
 const currentDateTitle = computed(() => {
   return `${currentMonth.value.getMonth() + 1}月`;
 });
@@ -224,7 +260,7 @@ const isError = computed(() => (form ? form.isError.value : false));
 
 const borderState = computed(() => {
   if (isError.value) return "error";
-  if (openDatepicker.value) return "active";
+  if (props.isOpen && !props.disabled) return "active";
   return "default";
 });