+++ packages/wiz-ui-next/src/components/base/popup/popup.vue
@@ -6,6 +6,7 @@
         shadow && popupShadowStyle,
         zIndexStyle[layer],
         !isActuallyOpen && popupHiddenStyle,
+        isFixed && popupFixedStyle,
       ]"
       :style="{
         inset,
@@ -30,6 +31,7 @@ import {
   popupStyle,
   popupShadowStyle,
   popupHiddenStyle,
+  popupFixedStyle,
 } from "@wizleap-inc/wiz-ui-styles/bases/popup.css";
 import { zIndexStyle } from "@wizleap-inc/wiz-ui-styles/commons";
 import {
@@ -157,6 +159,13 @@ const togglePopup = () => {
       isActuallyOpen.value = props.isOpen;
     };
   }
+};
+
+const existsFixedParent = (el: HTMLElement | null): HTMLElement | null => {
+  if (!el) return null;
+  const position = window.getComputedStyle(el).position;
+  if (position === "fixed") return el;
+  return existsFixedParent(el.parentElement);
 };
 
 let removeScrollHandler: (() => void) | null = null;
@@ -330,6 +339,10 @@ watch(
     emit("onTurn", newVal);
   }
 );
+
+const isFixed = computed(() => {
+  return existsFixedParent(containerRef.value || null) ? true : false;
+});
 
 const inset = computed(() => {
   const { scrollX, scrollY } = window;