+++ packages/wiz-ui-next/src/hooks/use-snackbar.ts
@@ -1,13 +1,12 @@
-import { h, render, ref, VNode } from "vue";
+import { ref, readonly } from "vue";
 
-import WizSnackbarController from "@/components/base/snackbar/snackbar-controller.vue";
 import { SnackbarOption } from "@/components/base/snackbar/types";
 
-const options = ref<SnackbarOption[]>([]);
-let snackbar: VNode | null = null;
+import { globalInject, globalKey } from "./use-global-provider";
 
-export const useSnackbar = () => {
-  const show = (message: string) => {
+export const useSnackbarManager = () => {
+  const options = ref<SnackbarOption[]>([]);
+  const snack = (message: string) => {
     const created = new Date().toISOString();
     const deleteSnackbar = () => {
       options.value = options.value.filter(
@@ -20,11 +19,16 @@ export const useSnackbar = () => {
       delete: deleteSnackbar,
     });
   };
+  return {
+    snackbarOptions: readonly(options),
+    snack,
+  };
+};
 
-  if (!snackbar) {
-    snackbar = h(WizSnackbarController, { options });
-    render(snackbar, document.body);
-  }
-
-  return show;
+export const useSnackbar = () => {
+  const { snack, snackbarOptions } = globalInject(globalKey);
+  return {
+    snack,
+    snackbarOptions,
+  };
 };