+++ packages/wiz-ui-react/src/components/base/inputs/text/components/index.ts
@@ -1,0 +1,1 @@
+export { WizTextInput } from "./text-input";