+++ packages/wiz-ui-react/src/components/base/tables/flat/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./flat-table";
+export * from "./flat-thead";
+export * from "./flat-tbody";
+export * from "./flat-tr";
+export * from "./flat-th";
+export * from "./flat-td";