+++ packages/wiz-ui-react/src/components/base/tables/unstyled/components/unstyled-td.tsx
@@ -1,0 +1,22 @@
+import { ComponentName } from "@wizleap-inc/wiz-ui-constants";
+import * as styles from "@wizleap-inc/wiz-ui-styles/bases/unstyled-table.css";
+import clsx from "clsx";
+import { ComponentProps, ReactNode } from "react";
+
+type Props = {
+  align?: "left" | "center" | "right";
+  children?: ReactNode;
+} & ComponentProps<"td">;
+
+const UnstyledTd = ({ align = "center", children, ...props }: Props) => (
+  <td
+    className={clsx(styles.unstyledTdStyle, styles.unstyledTdAlignStyle[align])}
+    {...props}
+  >
+    {children}
+  </td>
+);
+
+UnstyledTd.displayName = ComponentName.UnstyledTd;
+
+export const WizUnstyledTd = UnstyledTd;