+++ packages/wiz-ui/src/components/base/inputs/date-picker/date-picker.stories.ts
@@ -25,15 +25,37 @@ export default {
     input: {
       action: "input",
     },
+    disabled: {
+      control: {
+        type: "boolean",
+      },
+    },
+    isDirectionFixed: {
+      control: {
+        type: "boolean",
+      },
+    },
   },
 };
 
 const Template: StoryFn = (_, { argTypes }) => ({
   props: Object.keys(argTypes),
   components: { WizDatepicker, WizHStack },
+  setup() {
+    const date = ref<Date | null>(null);
+    const isOpen = ref(true);
+    const setIsOpen = (value: boolean) => (isOpen.value = value);
+    return { date, isOpen, setIsOpen };
+  },
   template: `
     <WizHStack>
-      <WizDatepicker v-bind="$props" @input="input"/>
+      <WizDatepicker 
+        v-bind="$props"
+        v-model="date"
+        :isOpen="isOpen"
+        @input="input"
+        @updateIsOpen="setIsOpen"
+      />
     </WizHStack>
   `,
 });
@@ -60,9 +82,12 @@ import { ref } from "vue";
 import { WizDatepicker } from "@wizleap-inc/wiz-ui";
 
 const date = ref<Date | null>(null);
+const input = (value: Date) => {
+  input.value = value;
+};
 </script>
 <template>
-  <WizDatepicker v-model="date" />
+  <WizDatepicker v-model="date" @input="input" />
 </template>
   `,
     },
@@ -86,9 +111,12 @@ import { ref } from "vue";
 import { WizDatepicker } from "@wizleap-inc/wiz-ui";
 
 const date = ref<Date | null>(null);
+const input = (value: Date) => {
+  input.value = value;
+};
 </script>
 <template>
-  <WizDatepicker v-model="date" placeholder="(例) 2000/1/1" />
+  <WizDatepicker v-model="date" @input="input" placeholder="(例) 2000/1/1" />
 </template>
   `,
     },
@@ -112,17 +140,45 @@ import { ref } from "vue";
 import { WizDatepicker } from "@wizleap-inc/wiz-ui";
 
 const date = ref<Date | null>(null);
+const input = (value: Date) => {
+  input.value = value;
+};
 </script>
 <template>
-  <WizDatepicker v-model="date" disabled />
+  <WizDatepicker v-model="date" @input="input" disabled />
 </template>
   `,
     },
   },
 };
 
+export const InitialValue: StoryFn<typeof WizDatepicker> = (
+  _,
+  { argTypes }
+) => ({
+  props: Object.keys(argTypes),
+  components: { WizDatepicker, WizHStack },
+  setup() {
+    const date = ref<Date | null>(new Date(1990, 0, 1));
+    const isOpen = ref(true);
+    const setIsOpen = (value: boolean) => (isOpen.value = value);
+    return { date, isOpen, setIsOpen };
+  },
+  template: `
+    <div>
+      <WizDatepicker 
+        v-bind="$props"
+        v-model="date"
+        :isOpen="isOpen"
+        @input="input"
+        @updateIsOpen="setIsOpen"
+      />
+    </div>
+  `,
+});
+
 const _formatDateSlash = (date: Date) => {
-  const year = date.getFullYear();
+  const year = date.getFullYear() % 100;
   const month = date.getMonth() + 1;
   const day = date.getDate();
   return `${year}/${month}/${day}`;
@@ -140,16 +196,53 @@ const _formatDateJpMonth = (date: Date) => {
   return `${month}月`;
 };
 
+export const Hover: StoryFn<typeof WizDatepicker> = (_, { argTypes }) => ({
+  props: Object.keys(argTypes),
+  components: { WizDatepicker, WizHStack },
+  setup() {
+    const date = ref<Date | null>(new Date(1990, 0, 1));
+    const isOpen = ref(true);
+    const isHover = ref(true);
+    const updateIsOpen = (value: boolean) => (isOpen.value = value);
+    const updateIsHover = (value: boolean) => (isHover.value = value);
+    return { date, isOpen, isHover, updateIsOpen, updateIsHover };
+  },
+  template: `
+    <div>
+      <WizDatepicker 
+        v-bind="$props"
+        v-model="date"
+        :isOpen="isOpen"
+        :isHover="isHover"
+        @input="input"
+        @updateIsOpen="updateIsOpen"
+        @updateIsHover="updateIsHover"
+      />
+    </div>
+  `,
+});
 export const Test: StoryFn<typeof WizDatepicker> = (_, { argTypes }) => ({
   props: Object.keys(argTypes),
   components: { WizDatepicker, WizHStack },
   setup() {
     const date = ref<Date | null>(null);
-    return { date };
+    const isOpen = ref(false);
+    const isHover = ref(false);
+    const updateIsOpen = (value: boolean) => (isOpen.value = value);
+    const updateIsHover = (value: boolean) => (isHover.value = value);
+    return { date, isOpen, isHover, updateIsOpen, updateIsHover };
   },
   template: `
     <div>
-      <WizDatepicker v-model="date" @input="input"/>
+      <WizDatepicker 
+        v-bind="$props"
+        v-model="date"
+        :isOpen="isOpen"
+        :isHover="isHover"
+        @input="input"
+        @updateIsOpen="updateIsOpen"
+        @updateIsHover="updateIsHover"
+      />
     </div>
   `,
 });
@@ -214,4 +307,30 @@ Test.play = async ({ canvasElement }) => {
     _formatDateJpMonth(clickDate)
   );
   await waitFor(() => expect(currentMonthDisplay).toBeTruthy());
+};
+
+export const IsDirectionFixed = Template.bind({});
+IsDirectionFixed.args = {
+  value: null,
+  isDirectionFixed: true,
+};
+IsDirectionFixed.parameters = {
+  docs: {
+    description: {
+      story: `isDirectionFixedを設定することで、Popup の表示位置を固定することができます。`,
+    },
+    source: {
+      code: `
+<script setup lang="ts">
+import { ref } from "vue";
+import { WizDatepicker } from "@wizleap-inc/wiz-ui";
+
+const date = ref<Date | null>(null);
+</script>
+<template>
+  <WizDatepicker v-model="date" isDirectionFixed />
+</template>
+  `,
+    },
+  },
 };