+++ packages/wiz-ui/src/components/base/inputs/date-range-picker/types.ts
@@ -1,7 +1,16 @@
-export interface DateRange {
-  start: Date | null;
-  end: Date | null;
-}
+export type DateRange =
+  | {
+      start: Date;
+      end: Date;
+    }
+  | {
+      start: Date;
+      end: null;
+    }
+  | {
+      start: null;
+      end: null;
+    };
 
 export interface DateRangePickerSelectBoxOption {
   label: string;